[local] isolate "template only nodeinfos provider"
And stop trying to make that a nodeinfo processor: this was an attempt
to follow upstream suggestions, but excessively intrusive (not rebase
friendly) for a feature we might keep localy/forked for a long time.

We're also submiting a "nodeinfos provider processor" to upstream, which
(if accepted) will help integrate that kind of changes much more cleanly
(eg. not breaking tests, not touching core/). For now, let's assume we
might not have an upstream  processor entry point for a long time.
