+++ cluster-autoscaler/core/static_autoscaler.go
@@ -48,6 +48,9 @@ import (
 	klog "k8s.io/klog/v2"
 )
 
+// GetNodeInfosForGroups is a non intrusive entrypoint for nodeinfos provider override
+var GetNodeInfosForGroups = core_utils.GetNodeInfosForGroups
+
 const (
 	// How old the oldest unschedulable pod should be before starting scale up.
 	unschedulablePodTimeBuffer = 2 * time.Second
@@ -267,18 +270,14 @@ func (a *StaticAutoscaler) RunOnce(currentTime time.Time) errors.AutoscalerError
 		return typedErr.AddPrefix("Initialize ClusterSnapshot")
 	}
 
-	var nodeInfosForGroups map[string]*schedulerframework.NodeInfo
-	var autoscalerError errors.AutoscalerError
-	if !a.AutoscalingContext.AutoscalingOptions.ScaleUpTemplateFromCloudProvider {
-		nodeInfosForGroups, autoscalerError = core_utils.GetNodeInfosForGroups(
-			readyNodes, a.nodeInfoCache, autoscalingContext.CloudProvider, autoscalingContext.ListerRegistry, daemonsets, autoscalingContext.PredicateChecker, a.ignoredTaints)
-		if autoscalerError != nil {
-			klog.Errorf("Failed to get node infos for groups: %v", autoscalerError)
-			return autoscalerError.AddPrefix("failed to build node infos for node groups: ")
-		}
+	nodeInfosForGroups, autoscalerError := GetNodeInfosForGroups(
+		readyNodes, a.nodeInfoCache, autoscalingContext.CloudProvider, autoscalingContext.ListerRegistry, daemonsets, autoscalingContext.PredicateChecker, a.ignoredTaints)
+	if autoscalerError != nil {
+		klog.Errorf("Failed to get node infos for groups: %v", autoscalerError)
+		return autoscalerError.AddPrefix("failed to build node infos for node groups: ")
 	}
 
-	nodeInfosForGroups, err = a.processors.NodeInfoProcessor.Process(autoscalingContext, nodeInfosForGroups, daemonsets, a.ignoredTaints)
+	nodeInfosForGroups, err = a.processors.NodeInfoProcessor.Process(autoscalingContext, nodeInfosForGroups)
 	if err != nil {
 		klog.Errorf("Failed to process nodeInfos: %v", err)
 		return errors.ToAutoscalerError(errors.InternalError, err)
