+++ cluster-autoscaler/processors/nodeinfos/node_info_processor.go
@@ -17,9 +17,6 @@ limitations under the License.
 package nodeinfos
 
 import (
-	appsv1 "k8s.io/api/apps/v1"
-	"k8s.io/autoscaler/cluster-autoscaler/utils/taints"
-
 	"k8s.io/autoscaler/cluster-autoscaler/context"
 	schedulerframework "k8s.io/kubernetes/pkg/scheduler/framework/v1alpha1"
 )
@@ -27,7 +24,7 @@ import (
 // NodeInfoProcessor processes nodeInfos after they're created.
 type NodeInfoProcessor interface {
 	// Process processes a map of nodeInfos for node groups.
-	Process(ctx *context.AutoscalingContext, nodeInfosForNodeGroups map[string]*schedulerframework.NodeInfo, daemonsets []*appsv1.DaemonSet, ignoredTaints taints.TaintKeySet) (map[string]*schedulerframework.NodeInfo, error)
+	Process(ctx *context.AutoscalingContext, nodeInfosForNodeGroups map[string]*schedulerframework.NodeInfo) (map[string]*schedulerframework.NodeInfo, error)
 	// CleanUp cleans up processor's internal structures.
 	CleanUp()
 }
@@ -37,7 +34,7 @@ type NoOpNodeInfoProcessor struct {
 }
 
 // Process returns unchanged nodeInfos.
-func (p *NoOpNodeInfoProcessor) Process(ctx *context.AutoscalingContext, nodeInfosForNodeGroups map[string]*schedulerframework.NodeInfo, daemonsets []*appsv1.DaemonSet, ignoredTaints taints.TaintKeySet) (map[string]*schedulerframework.NodeInfo, error) {
+func (p *NoOpNodeInfoProcessor) Process(ctx *context.AutoscalingContext, nodeInfosForNodeGroups map[string]*schedulerframework.NodeInfo) (map[string]*schedulerframework.NodeInfo, error) {
 	return nodeInfosForNodeGroups, nil
 }
